Split vtsearch/models into training/, embedding/, detectors/
Carry out step 2 of the codebase reorg's vtsearch/models split. The
package is gone; its contents are redistributed by concern:

- vtsearch/training/ — generic, media-agnostic learned-sort primitives:
  mlp.py (build_model, build_model_from_weights, train_model,
  _auto_hidden_dim), thresholds.py (GMM/cross-cal/safe/find-optimal +
  the cached cross-cal wrapper), svm.py (SVM trainer prototype),
  region_similarity.py (cosine-sim scoring with patch-region max-pool).
- vtsearch/embedding/ — embedder façades and torch runtime: helpers.py
  (embed_*_file / embed_text_query), loader.py (initialize_models,
  get_torch_device, preload_autoload_media_types, get_clap/xclip/e5
  getters), matrix.py (cached contiguous (N, D) embedding matrix).
- vtsearch/detectors/training.py — augmented with the vote-aware
  pipeline (_training_vec_for_vote, train_and_score) and the origin-
  based detector trainer (collect_media_origins,
  train_detector_from_origins). The pre-existing train_and_threshold /
  serialize_weights helpers stay alongside.
- vtsearch/detectors/workflow.py — was models/training_workflow.py
  (Flask-aware apply-labels-and-retrain).
- vtsearch/state/diversity_tree.py — was models/diversity_tree.py.

Updates ~50 import sites across the codebase, plus CLAUDE.md, ARCHITECTURE.md,
ML.md, EVAL.md, and the reorg plan to reflect the new layout. No
behaviour change — pure moves and a single training.py split.
