+++ CLAUDE.md
@@ -105,7 +105,9 @@ If a failure is genuinely outside the scope of the current task (e.g. a flaky ne
   - Top-level: `auth.py`, `eval.py`, `file_browser.py`, `main.py`, `sorting.py`, `achievements.py`
   - `_shared.py` — shared route helpers (request parsing, JSON safety, embedder lookup)
   - Note: dataset-load orchestration lives at `vtsearch/datasets/load_pipeline.py` (background-task helpers, ConcurrencyGate, clip fix-up).
-- `vtsearch/models/` — Embeddings, training, model loading, progress tracking, diversity tree
+- `vtsearch/detectors/` — Detector lifecycle and the resolve→embed→train pipeline. `registry.py` (in-memory detector registry), `store.py` (on-disk labelset/query store), `training.py` (vote-aware `train_and_score`, `train_and_threshold`, `train_detector_from_origins`, `collect_media_origins`), `workflow.py` (`apply_and_retrain` — Flask-aware), `resolver.py` (origin → file + embedding resolution), `label_sync.py`/`label_restoration.py`/`labelset_elements.py`/`labelset_training.py` (cross-dataset labelset materialisation + MLP training), `dataset_sync.py` (sync detectors when a dataset loads), `media_seeding.py` (seed training media), `labeling_progress.py` (per-step MLP cache + stability analysis)
+- `vtsearch/training/` — Generic learned-sort training primitives (media-agnostic). `mlp.py` (`build_model`, `build_model_from_weights`, `train_model`, `_auto_hidden_dim`), `thresholds.py` (`calculate_gmm_threshold`, `find_optimal_threshold`, `calculate_cross_calibration_threshold`, `cross_calibration_threshold_cached`, `calculate_safe_threshold`), `svm.py` (SVM trainer prototype), `region_similarity.py` (cosine-sim scoring with patch-region max-pool)
+- `vtsearch/embedding/` — Embedder façades and torch runtime. `helpers.py` (`embed_audio_file`/`embed_image_file`/`embed_video_file`/`embed_paragraph_file`/`embed_text_query` — thin wrappers around `MediaEmbedder`), `loader.py` (`initialize_models`, `preload_autoload_media_types`, `get_torch_device`, `get_clap_model`/`get_xclip_model`/`get_e5_model` getter wrappers), `matrix.py` (cached contiguous `(N, D)` embedding matrix on `DatasetContext`)
 - `vtsearch/datasets/` — Dataset loading, downloading, ingestion, origin tracking, labelsets, splitting. `loader.py` is the public façade that re-exports the actual loaders from sibling modules: `loader_folder.py` (folder loaders), `loader_pickle.py` (pickle loaders + sidecars + image embed), `loader_demo.py` (demo dataset loader). Importers live in `importers/` — `local_folder`/`local_files` (browser-side upload placeholders), `server_folder`/`server_files` (server filesystem paths), `pickle`, `http_archive`, `combine_datasets`, `demo`, `synthetic`, `recaller`; auto-discovered via the `IMPORTER` sentinel. `sources/` sub-package provides the `MediaSource` abstraction for resolving media files from origins (local_folder, http_archive, pullwrest)
 - `vtsearch/eval/` — Evaluation framework: runner, metrics, visualisation, voting iterations
 - `vtsearch/exporters/` — Results exporters (server_json_file/server_csv_file/email_smtp/webhook/gui/holder); auto-discovered via `EXPORTER` sentinel
@@ -117,7 +119,7 @@ If a failure is genuinely outside the scope of the current task (e.g. a flaky ne
 - `vtsearch/state/` — Multi-dataset / multi-detector global state. `core.py` owns `DatasetContext`, `DetectorContext`, `_state_lock`, the `medias`/votes proxy objects, and the context registries. Sub-modules split the operations by concern: `votes.py` (toggle_vote / apply_label / clear_votes), `clicks.py` (vote click times), `processors.py` (autorun extractor/localizer CRUD), `diversity.py` (diversity tree), `media_lookup.py` (origin-keyed lookup, collapse_duplicates). The package `__init__.py` re-exports every public name and hosts a handful of cross-cutting helpers (`snapshot_medias`, `clear_medias`, `clear_all`, `get_inclusion`/`set_inclusion`, dataset-display-name and safe-thresholds accessors)
 - `vtsearch/plugins/` — `PluginRegistry`, `PluginBase`, `PluginField`, and the sentinel-based auto-discovery scanner shared by every plugin family (dataset importers, exporters, label importers/sources, processor importers, settings importers/exporters/sources)
 - `vtsearch/sync/` — Generic `SyncSource[LoadT, SaveT]` base class (built on `PluginBase`) shared by `SettingsSource` and `LabelsetSource` for bidirectional load/save
-- `vtsearch/concurrency/` — Async job manager (`async_jobs.py` — `AsyncJob`, `JobManager`, `eval_jobs`, `learned_sort_jobs`), memory-aware worker capping (`memory_budget.py` — `cap_workers_by_memory`), and the long-running-operation progress infrastructure (`progress.py` — `ProgressTracker`, `LoadingTasksTracker`, `update_progress`/`get_progress` + sort/eval/find variants, `cancel_dataset_progress`, `check_dataset_cancelled`). The labeling-session analyzer in `vtsearch/models/labeling_progress.py` is a separate, unrelated module that caches per-step trained MLPs and computes stopping-condition metrics — the historical name collision with this package was resolved by renaming `models/progress.py`
+- `vtsearch/concurrency/` — Async job manager (`async_jobs.py` — `AsyncJob`, `JobManager`, `eval_jobs`, `learned_sort_jobs`), memory-aware worker capping (`memory_budget.py` — `cap_workers_by_memory`), and the long-running-operation progress infrastructure (`progress.py` — `ProgressTracker`, `LoadingTasksTracker`, `update_progress`/`get_progress` + sort/eval/find variants, `cancel_dataset_progress`, `check_dataset_cancelled`). The labeling-session analyzer in `vtsearch/detectors/labeling_progress.py` is a separate, unrelated module that caches per-step trained MLPs and computes stopping-condition metrics — the historical name collision with this package was resolved by renaming `models/progress.py` to `models/labeling_progress.py` (since moved into `vtsearch/detectors/`)
 - `vtsearch/security/` — Path / URL / pickle safety: `path_validation.py` (validate_server_filepath, sanitize_template_value, rglob_follow_symlinks, glob_top_level), `url_validation.py` (SSRF guard for outbound HTTP), `pickle.py` (safe_pickle_load + peek_pickle_dataset_summary for dataset pickles)
 - `vtsearch/utils/` — Leftover helpers that don't fit the packages above: `hits.py` (`build_media_hit` — the single source of truth for the scored-media hit dict shared by CLI autodetect and `/api/labels/fill-from-sort`), and `synthetic/` (offline media synthesis — `images.py`, `audio.py`, `video.py` for the SyntheticDatasetImporter)
 - `vtsearch/settings_factory.py` — Accessor factories (`make_accessors`, `make_per_side_setting`, `clamp`, `one_of`) used by `vtsearch/settings.py` to generate get/set pairs from the `_SETTING_SPECS` table
