+++ tests/fixtures/medias.py
@@ -10,7 +10,7 @@ from vtsearch.config import DATA_DIR
 from vtsearch.media.audio.media_type import generate_waveform_thumbnail
 
 NUM_MEDIAS = 20
-from vtsearch.models import embed_audio_file
+from vtsearch.embedding import embed_audio_file
 from vtsearch.state import medias
