+++ vtsearch/embedding/helpers.py
