+++ vtsearch/embedding/loader.py
