+++ vtsearch/state/diversity_tree.py
